MAILET-115 Remove abstract getTo implementation
